improve error handling

+++ source/blender/blenkernel/intern/bake_geometry_nodes_modifier_pack.cc
@@ -70,25 +70,37 @@ bool unpack_bake_to_disk(const NodesModifierPackedBake &packed_bake,
                          const BakePath &bake_path,
                          ReportList *reports)
 {
+  auto unpack_file = [&](const StringRefNull directory, const NodesModifierBakeFile &bake_file) {
+    char file_path[FILE_MAX];
+    BLI_path_join(file_path, sizeof(file_path), directory.c_str(), bake_file.name);
+    if (!BLI_file_ensure_parent_dir_exists(file_path)) {
+      BKE_reportf(reports, RPT_ERROR, "Can't ensure directory: %s", directory.c_str());
+      return false;
+    }
+    fstream fs(file_path, std::ios::out);
+    fs.write(static_cast<const char *>(bake_file.packed_file->data), bake_file.packed_file->size);
+    if (fs.bad()) {
+      BKE_reportf(reports, RPT_ERROR, "Can't write file : %s", file_path);
+      return false;
+    }
+    return true;
+  };
+
   for (const NodesModifierBakeFile &bake_file :
        Span{packed_bake.meta_files, packed_bake.meta_files_num})
   {
-    char file_path[FILE_MAX];
-    BLI_path_join(file_path, sizeof(file_path), bake_path.meta_dir.c_str(), bake_file.name);
-    BLI_file_ensure_parent_dir_exists(file_path);
-    fstream fs(file_path, std::ios::out);
-    fs.write(static_cast<const char *>(bake_file.packed_file->data), bake_file.packed_file->size);
+    if (!unpack_file(bake_path.meta_dir, bake_file)) {
+      return false;
+    }
   }
   for (const NodesModifierBakeFile &bake_file :
        Span{packed_bake.blob_files, packed_bake.blob_files_num})
   {
-    char file_path[FILE_MAX];
-    BLI_path_join(file_path, sizeof(file_path), bake_path.blobs_dir.c_str(), bake_file.name);
-    BLI_file_ensure_parent_dir_exists(file_path);
-    fstream fs(file_path, std::ios::out | std::ios::binary);
-    fs.write(static_cast<const char *>(bake_file.packed_file->data), bake_file.packed_file->size);
+    if (!unpack_file(bake_path.blobs_dir, bake_file)) {
+      return false;
+    }
   }
-  return false;
+  return true;
 }
 
 }  // namespace blender::bke::bake
